front end refactor to create project centric views
