+++ lib/javascript/utils/all.js
@@ -222,6 +222,28 @@ export function makeRequest(method, url, body, onprogressCallback, timeout) {
   });
 }
 
+export class PersistentLocalConfig {
+
+  constructor(localStoragePrefix){
+    this.localStoragePrefix = localStoragePrefix;
+  }
+
+  set(key, value){
+    return window.localStorage.setItem(`${this.localStoragePrefix}.${key}`, value);
+  }
+
+  get(key){
+    let value = window.localStorage.getItem(`${this.localStoragePrefix}.${key}`);
+    if(value === null){
+      value = undefined;
+    }
+    return value;
+  }
+  remove(key){
+    window.localStorage.removeItem(`${this.localStoragePrefix}.${key}`);
+  }
+}
+
 
 export function nodeCenter(el) {
