+++ services/orchest-webserver/app/app/templates/base.html
@@ -20,13 +20,17 @@
     <aside class="mdc-drawer mdc-drawer--dismissible" id="main-drawer">
         <div class="mdc-drawer__content">
             <nav class="mdc-list">
-                <a class="mdc-list-item mdc-list-item--activated" data-react-view="ProjectSelectionViewPipelines" href="#">
+                <a class="mdc-list-item mdc-list-item--activated" data-react-view="PipelinesView" href="#">
                     <i class="material-icons mdc-list-item__graphic" aria-hidden="true">device_hub</i>
                     <span class="mdc-list-item__text">Pipelines</span>
                 </a>
-                <a class="mdc-list-item" data-react-view="ProjectSelectionViewExperiments" href="#">
+                <a class="mdc-list-item" data-react-view="ExperimentsView" href="#">
                     <i class="material-icons mdc-list-item__graphic" aria-hidden="true">science</i>
                     <span class="mdc-list-item__text">Experiments</span>
+                </a>
+                <a class="mdc-list-item" data-react-view="ProjectsView" href="#">
+                    <i class="material-icons mdc-list-item__graphic" aria-hidden="true">format_list_bulleted</i>
+                    <span class="mdc-list-item__text">Projects</span>
                 </a>
                 <a class="mdc-list-item" data-react-view="FileManagerView" href="#">
                     <i class="material-icons mdc-list-item__graphic" aria-hidden="true">folder_open</i>
