+++ services/orchest-webserver/app/static/css/main.scss
@@ -101,7 +101,7 @@ h1, h2, h3 {
     }
   }
 
-  &.experiments-page {
+  .experiments-page {
     .mdc-data-table {
       width: 100%;
